pyproject.toml
[project]
name = "vacancy"
version = "0.1.0"
description = ""
authors = [
    {name = "giru",email = "yagofarov.vadim@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "pytest-django (>=4.9.0,<5.0.0)",
    "pytest-factoryboy (>=2.7.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
